fix: don't group by geometry on household_deficiency_afe_aggregation
The materialized view was taking over 50 minutes to refresh. This was
due to there being a clause on the view that grouped rows by
ST_AsGeoJSON(geometry), which was very computationally intensive.
Doing the group first and then joining the GeoJSON avoids Postgres
having to sort on a multi-megabyte field

#269